fix: power pad loc

+++ src/operation/iIR/source/module/power-netlist/PGNetlist.cc
@@ -500,7 +500,7 @@ void IRPGNetlistBuilder::build(
            << " bump node location: " << middle_point.get_x() / _dbu << " "
            << middle_point.get_y() / _dbu << " " << getLayerName(layer_id);
 
-  IRNodeLoc bump_node_loc{{middle_point.get_x() / _dbu, middle_point.get_y() / _dbu}, getLayerName(layer_id)};
+  IRNodeLoc bump_node_loc{{middle_point.get_x() / (double)_dbu, middle_point.get_y() / (double)_dbu}, getLayerName(layer_id)};
   _net_bump_node_locs[special_net_name] = bump_node_loc;
 
   LOG_INFO << "instance pin edge num: "
